# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Headway API",
    "description": "Serves computed transit metrics with full lineage, the DQ resolution workflow, and the audited certification action. This API never computes a reported figure; it serves what the calculation library produced, joined to its provenance. Reported values are JSON strings (exact NUMERIC, never float). Auth: local accounts (ADR-0011); the native OIDC relying party is the next increment behind the same {sub, username, role} claim set.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/dq/issues": {
      "get": {
        "tags": [
          "data-quality"
        ],
        "summary": "List Issues",
        "operationId": "list_issues_dq_issues_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DqIssue"
                  },
                  "title": "Response List Issues Dq Issues Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dq/issues/counts": {
      "get": {
        "tags": [
          "data-quality"
        ],
        "summary": "Count Issues",
        "description": "Severity/status counts over the same rows (and the same optional\nstatus filter) as GET /dq/issues \u2014 composition of the one issues query,\ncounted in the open, no new tables (handoff 0017).",
        "operationId": "count_issues_dq_issues_counts_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DqIssueCounts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dq/issues/{issue_id}/resolve": {
      "post": {
        "tags": [
          "data-quality"
        ],
        "summary": "Resolve Issue",
        "operationId": "resolve_issue_dq_issues__issue_id__resolve_post",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/reports/mr20": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Mr20 Report",
        "description": "Serve the MR-20 preview package (NOT reportable) for one month.",
        "operationId": "get_mr20_report_reports_mr20_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Calendar month as YYYY-MM (e.g. 2026-07), UTC, half-open period.",
              "title": "Month"
            },
            "description": "Calendar month as YYYY-MM (e.g. 2026-07), UTC, half-open period."
          }
        ],
        "responses": {
          "200": {
            "description": "The headway_calc.mr20 package for the month, verbatim: form/generator/period header, reportable=false + banner, programmatically enumerated caveats, and the four MR-20 data points per mode plus fleet totals, each cell carrying full provenance or an explicit null + reason.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/mr20/export": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Export Mr20 Report",
        "description": "Download the MR-20 preview package as CSV/XLSX (handoff 0017, design\npoint 5). Same package build as GET /reports/mr20; one row assembly\nfeeds both formats (XLSX values byte-equal to CSV values, pinned by\ntest); XLSX cells are TEXT so figures survive exactly.",
        "operationId": "export_mr20_report_reports_mr20_export_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Calendar month as YYYY-MM.",
              "title": "Month"
            },
            "description": "Calendar month as YYYY-MM."
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(csv|xlsx)$",
              "default": "xlsx",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The MR-20 preview package as a CSV or XLSX download: one row per (scope, metric) cell, values VERBATIM from the package; the NOT-REPORTABLE banner and every enumerated caveat lead the CSV and form the XLSX's first sheet.",
            "content": {
              "text/csv; charset=utf-8": {},
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/ss50/export": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Export Ss50 Report",
        "description": "Download the S&S-50 preview package (headway_calc.ss50, served\nverbatim into a grid) as CSV/XLSX \u2014 handoff 0017, design point 5.",
        "operationId": "export_ss50_report_reports_ss50_export_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Calendar month as YYYY-MM.",
              "title": "Month"
            },
            "description": "Calendar month as YYYY-MM."
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(csv|xlsx)$",
              "default": "xlsx",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The S&S-50 non-major monthly summary package as a CSV or XLSX download: one row per (mode, type-of-service) cell including explicit zero rows; the NOT-REPORTABLE banner, citations, caveats and the excluded-event accounting lead the CSV and form the XLSX's first sheet.",
            "content": {
              "text/csv; charset=utf-8": {},
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DqIssue": {
        "properties": {
          "issue_id": {
            "type": "string",
            "title": "Issue Id"
          },
          "issue_type": {
            "type": "string",
            "title": "Issue Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "source_record_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Record Ids"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "resolution_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Minutes"
          }
        },
        "type": "object",
        "required": [
          "issue_id",
          "issue_type",
          "severity",
          "status",
          "owner",
          "title",
          "description",
          "source_record_ids",
          "created_at",
          "resolved_at",
          "resolution",
          "resolution_minutes"
        ],
        "title": "DqIssue"
      },
      "DqIssueCounts": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Severity"
          },
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Status"
          }
        },
        "type": "object",
        "required": [
          "total",
          "by_severity",
          "by_status"
        ],
        "title": "DqIssueCounts",
        "description": "Counts for the /dq summary cards (handoff 0017, design point 2):\ncounted over EXACTLY the rows GET /dq/issues serves under the same\nfilter, so a card total can never disagree with the table below it.\nMissing severities/statuses appear as explicit zeros."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolveRequest": {
        "properties": {
          "resolution": {
            "type": "string",
            "minLength": 1,
            "title": "Resolution"
          },
          "resolution_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Minutes"
          }
        },
        "type": "object",
        "required": [
          "resolution"
        ],
        "title": "ResolveRequest"
      },
      "ResolveResponse": {
        "properties": {
          "issue_id": {
            "type": "string",
            "title": "Issue Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "resolved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Resolved At"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution"
          },
          "resolution_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Minutes"
          },
          "audit_event_id": {
            "type": "integer",
            "title": "Audit Event Id"
          }
        },
        "type": "object",
        "required": [
          "issue_id",
          "status",
          "resolved_at",
          "resolution",
          "resolution_minutes",
          "audit_event_id"
        ],
        "title": "ResolveResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}